+++ plugins/content/content_quote_doer/quote.py
@@ -1,4 +1,4 @@
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 from skevk import upload_photo
 from utils import traverse