+++ plugins/custom/__init__.py
@@ -1,0 +1,1 @@
+"""Pack with your custom plugins"""