+++ plugins/outsource/outsource_audio2text.py
@@ -1,4 +1,4 @@
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 from utils import traverse
 
 import uuid, defusedxml.ElementTree as ET